util added

+++ storage-service/src/main/java/com/nexis/storage_service/dto/FileRequestDto.java
@@ -6,6 +6,6 @@ import java.util.UUID;
 public record FileRequestDto(
         UUID workspaceId,
         String fileName,
-        BigInteger size
+        Long size
 ) {
 }
