+++ storage-service/src/main/java/com/nexis/storage_service/entity/FileEntity.java
@@ -16,7 +16,7 @@ public class FileEntity {
     private UUID id;
     private UUID workspaceId;
     private String fileName;
-    private BigInteger fileSize;
+    private Long fileSize;
     private String fileType;
     private Integer currentVersion;
     private String storageKey;
